Changed the handling of run() and the hiding of steps.
The return value of run is not checked anymore
When something in run fails, YOU HAVE TO raise, otherwise it will
continue on its merry way.

A step has a should_run function that can be used to hide itself
from the steps when it does nothing. Only use should_run when it
makes no sense to show the step. A step like Fetch::HTTP should
still be shown even when the file already exists.